requirements.txt
appdirs==1.4.4
black==20.8b1
Brotli==1.0.9
click==7.1.2
dash==1.16.3
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
flake8==3.8.4
Flask==1.1.2
Flask-Compress==1.7.0
future==0.18.2
gunicorn==20.0.4
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
numpy==1.19.2
pandas==1.1.3
pathspec==0.8.0
plotly==4.11.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.10.11
retrying==1.3.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.3.0
